+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Release Notes
 
+`1.0.11`
+
+- Added `Methods.getCallerMethodName` without a supplier to get the caller method name.
+
 ---
 
 `1.0.10`